Add ToXNumber and make ToXBool work properly for types with length
