+++ excellent/types/conversions_test.go
@@ -9,7 +9,200 @@ import (
 	"github.com/nyaruka/goflow/utils"
 
 	"github.com/shopspring/decimal"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
+
+type TestXObject struct {
+	foo string
+	bar int
+}
+
+func NewTestXObject(foo string, bar int) *TestXObject {
+	return &TestXObject{foo: foo, bar: bar}
+}
+
+// ToJSON converts this type to JSON
+func (v *TestXObject) ToJSON() types.XString {
+	e := struct {
+		Foo string `json:"foo"`
+		Bar int    `json:"bar"`
+	}{
+		Foo: v.foo,
+		Bar: v.bar,
+	}
+	return types.RequireMarshalToXString(e)
+}
+
+func (v *TestXObject) Reduce() types.XPrimitive { return types.NewXString(v.foo) }
+
+var _ types.XValue = &TestXObject{}
+
+func TestXValueRequiredConversions(t *testing.T) {
+	chi, err := time.LoadLocation("America/Chicago")
+	require.NoError(t, err)
+
+	date1 := time.Date(2017, 6, 23, 15, 30, 0, 0, time.UTC)
+	date2 := time.Date(2017, 7, 18, 15, 30, 0, 0, chi)
+
+	tests := []struct {
+		value    types.XValue
+		asJSON   string
+		asString string
+		asBool   bool
+	}{
+		{
+			value:    types.NewXString(""),
+			asJSON:   `""`,
+			asString: "",
+			asBool:   false, // empty strings are false
+		}, {
+			value:    types.NewXString("FALSE"),
+			asJSON:   `"FALSE"`,
+			asString: "FALSE",
+			asBool:   false, // because it's string value is "false"
+		}, {
+			value:    types.NewXString("hello \"bob\""),
+			asJSON:   `"hello \"bob\""`,
+			asString: "hello \"bob\"",
+			asBool:   true,
+		}, {
+			value:    types.NewXNumberFromInt(0),
+			asJSON:   `0`,
+			asString: "0",
+			asBool:   false, // because any decimal != 0 is true
+		}, {
+			value:    types.NewXNumberFromInt(123),
+			asJSON:   `123`,
+			asString: "123",
+			asBool:   true, // because any decimal != 0 is true
+		}, {
+			value:    types.RequireXNumberFromString("123.00"),
+			asJSON:   `123`,
+			asString: "123",
+			asBool:   true,
+		}, {
+			value:    types.RequireXNumberFromString("123.45"),
+			asJSON:   `123.45`,
+			asString: "123.45",
+			asBool:   true,
+		}, {
+			value:    types.NewXBool(false),
+			asJSON:   `false`,
+			asString: "false",
+			asBool:   false,
+		}, {
+			value:    types.NewXBool(true),
+			asJSON:   `true`,
+			asString: "true",
+			asBool:   true,
+		}, {
+			value:    types.NewXTime(date1),
+			asJSON:   `"2017-06-23T15:30:00.000000Z"`,
+			asString: "2017-06-23T15:30:00.000000Z",
+			asBool:   true,
+		}, {
+			value:    types.NewXTime(date2),
+			asJSON:   `"2017-07-18T15:30:00.000000-05:00"`,
+			asString: "2017-07-18T15:30:00.000000-05:00",
+			asBool:   true,
+		}, {
+			value:    types.NewXError(fmt.Errorf("it failed")),
+			asJSON:   `"it failed"`,
+			asString: "it failed",
+			asBool:   false,
+		}, {
+			value:    types.NewXArray(),
+			asJSON:   `[]`,
+			asString: `[]`,
+			asBool:   false,
+		}, {
+			value:    types.NewXArray(types.NewXTime(date1), types.NewXTime(date2)),
+			asJSON:   `["2017-06-23T15:30:00.000000Z","2017-07-18T15:30:00.000000-05:00"]`,
+			asString: `["2017-06-23T15:30:00.000000Z","2017-07-18T15:30:00.000000-05:00"]`,
+			asBool:   true,
+		}, {
+			value:    NewTestXObject("Hello", 123),
+			asJSON:   `{"foo":"Hello","bar":123}`,
+			asString: "Hello",
+			asBool:   true,
+		}, {
+			value:    NewTestXObject("", 123),
+			asJSON:   `{"foo":"","bar":123}`,
+			asString: "",
+			asBool:   false, // because it reduces to a string which itself is false
+		}, {
+			value:    types.NewXArray(NewTestXObject("Hello", 123), NewTestXObject("World", 456)),
+			asJSON:   `[{"foo":"Hello","bar":123},{"foo":"World","bar":456}]`,
+			asString: `["Hello","World"]`,
+			asBool:   true,
+		}, {
+			value: types.NewXMap(map[string]types.XValue{
+				"first":  NewTestXObject("Hello", 123),
+				"second": NewTestXObject("World", 456),
+			}),
+			asJSON:   `{"first":{"foo":"Hello","bar":123},"second":{"foo":"World","bar":456}}`,
+			asString: `{"first":"Hello","second":"World"}`,
+			asBool:   true,
+		}, {
+			value:    types.NewXJSONArray([]byte(`[]`)),
+			asJSON:   `[]`,
+			asString: `[]`,
+			asBool:   false,
+		}, {
+			value:    types.NewXJSONArray([]byte(`[5, "x"]`)),
+			asJSON:   `[5, "x"]`,
+			asString: `[5, "x"]`,
+			asBool:   true,
+		}, {
+			value:    types.NewXJSONObject([]byte(`{}`)),
+			asJSON:   `{}`,
+			asString: `{}`,
+			asBool:   false,
+		}, {
+			value:    types.NewXJSONObject([]byte(`{"foo":"World","bar":456}`)),
+			asJSON:   `{"foo":"World","bar":456}`,
+			asString: `{"foo":"World","bar":456}`,
+			asBool:   true,
+		},
+	}
+	for _, test := range tests {
+		assert.Equal(t, types.NewXString(test.asJSON), types.ToXJSON(test.value), "ToXJSON failed for %+v", test.value)
+		assert.Equal(t, types.NewXString(test.asString), types.ToXString(test.value), "ToXString failed for %+v", test.value)
+		assert.Equal(t, types.NewXBool(test.asBool), types.ToXBool(test.value), "ToXBool failed for %+v", test.value)
+	}
+}
+
+func TestToXNumber(t *testing.T) {
+	//testResolver := &resolver{"155"}
+
+	var tests = []struct {
+		value    types.XValue
+		asNumber types.XNumber
+		hasError bool
+	}{
+		{nil, types.XNumberZero, false},
+		{types.NewXError(fmt.Errorf("Error")), types.XNumberZero, true},
+		{types.NewXNumberFromInt(123), types.NewXNumberFromInt(123), false},
+		{types.NewXString("15.5"), types.RequireXNumberFromString("15.5"), false},
+		{types.NewXString("lO.5"), types.RequireXNumberFromString("10.5"), false},
+		{NewTestXObject("Hello", 123), types.XNumberZero, true},
+		{NewTestXObject("123.45000", 123), types.RequireXNumberFromString("123.45"), false},
+	}
+
+	for _, test := range tests {
+		result, err := types.ToXNumber(test.value)
+
+		if test.hasError {
+			assert.Error(t, err, "expected error for input '%s'", test.value)
+		} else {
+			assert.NoError(t, err, "unexpected error for input '%s'", test.value)
+			assert.Equal(t, test.asNumber.Native(), result.Native(), "result mismatch for input '%+v'", test.value)
+		}
+	}
+}
+
+// Legacy...
 
 // test variable resolver
 type resolver struct {
